+++ client/src/socket/socketSubscriptions.js
@@ -1,26 +1,3 @@
-import io from 'socket.io-client';
-
-export const initSocket = ({ url, config = {} }) => {
-  const socket = io(url, config);
-
-  // Needs to be able to pass a callback
-  socket.on('connect', () => {
-    console.log('connected');
-  });
-
-  // placeholder event
-  socket.on('event', () => {
-    console.log('event');
-  });
-
-  // Needs to be able to pass a callback
-  socket.on('disconnect', () => {
-    console.log('disconnected');
-  });
-
-  return socket;
-};
-
 export const subscribeToClientRegister = (socket, callback) => {
   socket.on('register-client', uniqueID => callback(uniqueID));
 };
@@ -42,9 +19,3 @@ export const subscribeToPlayerRegister = (socket, callback) => {
 // export const subscribeTo = (socket, callback) => {
 //   socket.on('eventTosubscribeTo', data => callback(data));
 // };
-
-export const socketEmitter = (socket, event, data) => {
-  console.log(event, data);
-
-  socket.emit(event, data);
-};
